First approach of time coherence
Here, compute the mean between the previous mean and the current one. However, the results are not quite good.
This is because the new position has the half weight that the mean of the previous frame. Then, it changes a lot the mean value that will be get as center of the bbox.
In order to improve, we are going to build a history array that will keep the values of all the previous position, and we are going to obtain the mean at each iteration using this. Also, we are going to compute the distance between the new frame bbox and the previous one. This will be done in the branc time-coherence-option-2